Rewriting database management using diesel.
The QueryBuilder and its implementation were removed. I created migrations to be performed with diesel-cli, as well as models for each operation to be performed on the table in question. I created a migration system from SQLite to PostgreSQL. I also created a Docker Compose file to have a local PostgreSQL database and an .env file with all the system configuration.